Fix doc examples to use DeriveDataTypeable language extension
Use of LANGUAGE DeriveDataTypeable is required in order to compile the
second example.

Add OverloadedStrings and get rid of the "encodeStr" call.  I'd
expect most users of this library to use it with OverloadedStrings, so
I think it's nice for the "advanced" example to show this usage.

Also align import columns.