Generic context creation for `Maybe` (fixes #38)
The old behaviour was due to the following fact: `mkGenericContext` uses
SYB to traverse the datatype and to add everything it can convert to
`MuValue` to the context. The generated context would return `MuNothing`
if queried regarding a field that is not present, or a field that was
not processed. Since `Maybe` types were not supported, all the queries
regarding them were resulting in `MuNothing`.

Now, the `Maybe` constructor in generic contexts seem to be behaving
correctly. See `nestedPolyGenericContextTest` in the testsuite.